requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
gunicorn==20.1.0
img2pdf==0.4.3
lxml==4.7.1
packaging==21.3
pikepdf==4.3.1
Pillow==9.0.0
psycopg2==2.9.3
pyparsing==3.0.7
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
